move most of branch functionality to trait
ARD methods are still wrong, need to be adapted to
multiple precisions
